# backend/cmd/order-service/docs/swagger.yaml
basePath: /
definitions:
...
  handler.GatewayChangePasswordBody:
    properties:
      new_password:
        example: newS3cret!
        type: string
      old_password:
        example: s3cret!Pass
        type: string
    type: object
...
  handler.GatewayErrorResponse:
    properties:
      error:
        example: service unavailable
        type: string
    type: object
...
  handler.GatewayLoginBody:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: s3cret!Pass
        type: string
    type: object
  handler.GatewayLogoutBody:
    properties:
      refresh_token:
        example: eyJhbGci...
        type: string
    type: object
...
  handler.GatewayRefreshBody:
    properties:
      refresh_token:
        example: eyJhbGci...
        type: string
    type: object
  handler.GatewayRegisterBody:
    properties:
      email:
        example: alice@example.com
        type: string
      name:
        example: Alice Smith
        type: string
      password:
        example: s3cret!Pass
        type: string
    type: object
  handler.GatewayTokenPairResponse:
    properties:
      access_token:
        example: eyJhbGci...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGci...
        type: string
    type: object
  handler.GatewayUserResponse:
    properties:
      created_at:
        example: "2026-03-13T00:00:00Z"
        type: string
      email:
        example: alice@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Alice Smith
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Manages order creation, retrieval, and status updates driven by payment
    events.
  title: Order Service
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password and revokes all existing
        refresh tokens.
      parameters:
      - description: Old and new passwords
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayChangePasswordBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns a token pair. Use the access_token
        as "Bearer <token>" in the Authorize dialog.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayLoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GatewayTokenPairResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the given refresh token. Access tokens expire naturally
        after 15 min.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayLogoutBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GatewayUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new token pair. The old refresh
        token is revoked immediately.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayRefreshBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GatewayTokenPairResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns an access token (15 min)
        + refresh token (7 days).
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayRegisterBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GatewayTokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
        "409":
          description: Email already taken
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      summary: Register
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: JWT token issued by the API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"